requirements.txt
asgiref==3.8.1
certifi==2025.1.31
coverage==7.4.1
distlib==0.3.9
Django==4.2.10
django-cors-headers==4.3.1
django-filter==23.5
django-mssql==1.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
filelock==3.17.0
inflection==0.5.1
mssql-django==1.5
packaging==24.2
Pillow==10.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.10.1
pyodbc==5.2.0
python-dotenv==0.21.1
python-env==1.0.0
pytz==2025.1
PyYAML==6.0.2
setuptools==76.0.0
sqlparse==0.5.3
uritemplate==4.1.1
virtualenv==20.29.2
